webApp/requirements.txt
PyYAML==6.0.1
requests==2.31.0
sendgrid==6.11.0
Jinja2==3.1.3
